+++ pkg/deployment/resources/inspector/at_anonymous_v1.go
@@ -24,12 +24,47 @@ import (
 	"context"
 
 	meta "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/types"
+
+	api "github.com/arangodb/kube-arangodb/pkg/apis/deployment/v1"
 )
 
 type arangoTasksInspectorAnonymousV1 struct {
-	i *arangoTasksInspectorV1
+	i *inspectorState
 }
 
 func (e *arangoTasksInspectorAnonymousV1) Get(ctx context.Context, name string, opts meta.GetOptions) (meta.Object, error) {
-	return e.i.Get(ctx, name, opts)
+	return e.i.arangoTasks.v1.Get(ctx, name, opts)
+}
+
+func (e *arangoTasksInspectorAnonymousV1) Create(ctx context.Context, obj meta.Object, opts meta.CreateOptions) (meta.Object, error) {
+	if o, ok := obj.(*api.ArangoTask); !ok {
+		return nil, newInvalidTypeError(ArangoTaskGKv1())
+	} else {
+		return e.i.ArangoTaskModInterface().V1().Create(ctx, o, opts)
+	}
+}
+
+func (e *arangoTasksInspectorAnonymousV1) Update(ctx context.Context, obj meta.Object, opts meta.UpdateOptions) (meta.Object, error) {
+	if o, ok := obj.(*api.ArangoTask); !ok {
+		return nil, newInvalidTypeError(ArangoTaskGKv1())
+	} else {
+		return e.i.ArangoTaskModInterface().V1().Update(ctx, o, opts)
+	}
+}
+
+func (e *arangoTasksInspectorAnonymousV1) UpdateStatus(ctx context.Context, obj meta.Object, opts meta.UpdateOptions) (meta.Object, error) {
+	if o, ok := obj.(*api.ArangoTask); !ok {
+		return nil, newInvalidTypeError(ArangoTaskGKv1())
+	} else {
+		return e.i.ArangoTaskModInterface().V1().UpdateStatus(ctx, o, opts)
+	}
+}
+
+func (e *arangoTasksInspectorAnonymousV1) Patch(ctx context.Context, name string, pt types.PatchType, data []byte, opts meta.PatchOptions, subresources ...string) (result meta.Object, err error) {
+	return e.i.ArangoTaskModInterface().V1().Patch(ctx, name, pt, data, opts, subresources...)
+}
+
+func (e *arangoTasksInspectorAnonymousV1) Delete(ctx context.Context, name string, opts meta.DeleteOptions) error {
+	return e.i.ArangoTaskModInterface().V1().Delete(ctx, name, opts)
 }