+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - (Improvement) Improve error message when replication can't be configured
 - (Bugfix) Fix License handling in case of broken license secret
 - (Bugfix) Check ArangoSync availability without checking healthiness
+- (Improvement) Add Anonymous Inspector mods
 
 ## [1.2.20](https://github.com/arangodb/kube-arangodb/tree/1.2.20) (2022-10-25)
 - (Feature) Add action progress